Convert notification redux actions to TS.
